Fix merge issues
Fix merge issues due to reset context addition in the
AnnotationEventHandlerAdapter and their respective tests

#901
